+++ nova/modules/core/config/info.php
@@ -19,7 +19,7 @@ define('SMS_UPGRADE_VERSION', '2.6.9');
 //define('VERSION_FEED', MODPATH.'assets/version.yml');
 define('VERSION_FEED', 'https://anodyne-productions.com/feeds/version_nova.yml');
 
-define('REGISTER', 'https://anodyne-productions.com/register/index.php/utility/do_registration');
+define('REGISTER_URL', 'https://admin.anodyne-productions.com/api/games');
 
 // figure out whether to install the bare essentials or the dev stuff
 define('APP_DATA_SRC', 'basic');